+++ packages/flutter_meedu/test/router/navigation_test.dart
@@ -8,28 +8,37 @@ void main() {
   });
   group('navigation', () {
     testWidgets('normal navigation', (test) async {
-      router.setDefaultTransition(router.Transition.material,
-          duration: Duration(milliseconds: 100));
+      router.setDefaultTransition(
+        router.Transition.material,
+        duration: Duration(milliseconds: 100),
+      );
       await test.pumpWidget(
         MaterialApp(
           navigatorKey: router.navigatorKey,
           home: HomePage(),
+          navigatorObservers: [
+            router.observer,
+          ],
           routes: {
             'detail': (_) => DetailPage(),
           },
         ),
       );
       expect(router.canPop(), false);
+      expect(router.arguments, isNull);
       await test.tap(find.text("pushNamed"));
       await test.pumpAndSettle();
+      expect(router.arguments, "meedu");
       expect(find.text("meedu"), findsOneWidget);
       expect(router.canPop(), true);
       await test.tap(find.text("back"));
       await test.pumpAndSettle();
       expect(find.text("meedu"), findsNothing);
+      expect(router.arguments, isNull);
       await test.tap(find.text("pushReplacement"));
       await test.pumpAndSettle();
       expect(find.text("meedu"), findsOneWidget);
+      expect(router.arguments, "meedu");
       expect(router.canPop(), false);
     });
   });
@@ -41,6 +50,9 @@ void main() {
         key: router.appKey,
         initialRoute: '/',
         navigatorKey: router.navigatorKey,
+        navigatorObservers: [
+          router.observer,
+        ],
         routes: {
           '/': (_) => HomePage(),
           '/detail': (_) => DetailPage(),
@@ -51,6 +63,36 @@ void main() {
     await tester.pumpAndSettle();
     expect(find.text("meedu"), findsOneWidget);
     expect(find.text("popAndPushNamed"), findsNothing);
+  });
+
+  testWidgets('initial arguments', (tester) async {
+    router.setDefaultTransition(router.Transition.downToUp);
+    await tester.pumpWidget(
+      MaterialApp(
+        key: router.appKey,
+        onGenerateInitialRoutes: (String name) {
+          return [
+            MaterialPageRoute(
+              builder: (_) => DetailPage(),
+              settings: RouteSettings(
+                name: name,
+                arguments: "meedu",
+              ),
+            ),
+          ];
+        },
+        navigatorKey: router.navigatorKey,
+        initialRoute: '/',
+        navigatorObservers: [
+          router.observer,
+        ],
+        routes: {
+          '/demo': (_) => Container(),
+        },
+      ),
+    );
+
+    expect(find.text("meedu"), findsOneWidget);
   });
 }
 
@@ -97,7 +139,7 @@ class HomePage extends StatelessWidget {
 class DetailPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final productName = router.arguments<String>(context);
+    final productName = router.arguments as String;
     return Scaffold(
       appBar: AppBar(
         leading: TextButton(